requirements.txt
streamlit
google-generativeai
python-dotenv
folium
streamlit-folium
altair                       ==5.4.1
annotated-types              ==0.7.0
asttokens                    ==2.4.1
attrs                        ==24.2.0
blinker                      ==1.8.2
cachetools                   ==5.5.0
certifi                      ==2024.8.30
charset-normalizer           ==3.3.2
click                        ==8.1.7
colorama                     ==0.4.6
decorator                    ==5.1.1
executing                    ==2.1.0
gitdb                        ==4.0.11
GitPython                    ==3.1.43
google-ai-generativelanguage ==0.6.10
google-api-core              ==2.20.0
google-api-python-client     ==2.147.0
google-auth                  ==2.35.0
google-auth-httplib2         ==0.2.0
google-generativeai          ==0.8.2
googleapis-common-protos     ==1.65.0
grpcio                       ==1.66.1
grpcio-status                ==1.66.1
httplib2                     ==0.22.0
idna                         ==3.10
ipython                      ==8.27.0
jedi                         ==0.19.1
Jinja2                       ==3.1.4
jsonschema                   ==4.23.0
jsonschema-specifications    ==2023.12.1
markdown-it-py               ==3.0.0
MarkupSafe                   ==2.1.5
matplotlib-inline            ==0.1.7
mdurl                        ==0.1.2
narwhals                     ==1.8.3
numpy                        ==2.1.1
packaging                    ==24.1
pandas                       ==2.2.3
parso                        ==0.8.4
pillow                       ==10.4.0
pip                          ==24.2
prompt_toolkit               ==3.0.48
proto-plus                   ==1.24.0
protobuf                     ==5.28.2
pure_eval                    ==0.2.3
pyarrow                      ==17.0.0
pyasn1                       ==0.6.1
pyasn1_modules               ==0.4.1
pydantic                     ==2.9.2
pydantic_core                ==2.23.4
pydeck                       ==0.9.1
Pygments                     ==2.18.0
pyparsing                    ==3.1.4
python-dateutil              ==2.9.0.post0
pytz                         ==2024.2
referencing                  ==0.35.1
requests                     ==2.32.3
rich                         ==13.8.1
rpds-py                      ==0.20.0
rsa                          ==4.9
six                          ==1.16.0
smmap                        ==5.0.1
stack-data                   ==0.6.3
streamlit                    ==1.38.0
tenacity                     ==8.5.0
toml                         ==0.10.2
tornado                      ==6.4.1
tqdm                         ==4.66.5
traitlets                    ==5.14.3
typing_extensions            ==4.12.2
tzdata                       ==2024.2
uritemplate                  ==4.1.1
urllib3                      ==2.2.3
watchdog                     ==4.0.2
wcwidth                      ==0.2.13
PyPDF2==3.0.1
python-docx==0.8.11
docx2pdf==0.1.8
reportlab
streamlit-elements==0.1.*
plotly
openpyxl
matplotlib
mediapipe
opencv-python